backend/requirements.txt

paho-mqtt==1.6.1
flask==3.0.0
flask-cors==4.0.0